[LexicalLookup] Make `GenericParameterScopeSyntax` public
Using a method in a extension of non-public protocol to witness a
public protocol requirement is not allowed in future version of Swift
and is diagnosed as a warning in the current main.

Make `GenericParameterScopeSyntax` a `@_spi(Experimetnal public)` the
same accessibility as the concrete type.

(cherry picked from commit 265509f5ea49401598b9741b442f9c82d6240c5b)